+++ app/src/main/java/com/example/lifesync/StepCountCallback.java
@@ -7,11 +7,11 @@ public interface StepCountCallback {
 
 class SensorData {
     final int steps;
-    final int bmi;
+    final double bmi;
     final double distance;
-    final int calories;
+    final double calories;
 
-    SensorData(int steps, int bmi, double distance, int calories) {
+    SensorData(int steps, double bmi, double distance, double calories) {
         this.steps = steps;
         this.bmi = bmi;
         this.distance = distance;
